+++ common/src/main/java/de/maxhenkel/voicechat/gui/SelectMicrophoneScreen.java
@@ -11,7 +11,6 @@ import net.minecraft.ChatFormatting;
 import net.minecraft.client.gui.components.Button;
 import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.network.chat.Component;
-import net.minecraft.network.chat.MutableComponent;
 import net.minecraft.network.chat.TextComponent;
 import net.minecraft.network.chat.TranslatableComponent;
 
@@ -19,10 +18,14 @@ import javax.annotation.Nullable;
 
 public class SelectMicrophoneScreen extends ListScreen<String> {
 
+    private static final Component TITLE = new TranslatableComponent("gui.voicechat.select_microphone.title");
+    private static final Component SELECT = new TranslatableComponent("message.voicechat.select");
+    private static final Component NO_MICROPHONE = new TranslatableComponent("message.voicechat.no_microphone").withStyle(ChatFormatting.WHITE);
+
     protected int selected;
 
     public SelectMicrophoneScreen(Screen parent) {
-        super(parent, MicrophoneManager.deviceNames(), new TranslatableComponent("gui.voicechat.select_microphone.title"));
+        super(parent, MicrophoneManager.deviceNames(), TITLE);
         for (int i = 0; i < elements.size(); i++) {
             String element = elements.get(i);
             if (element.equals(VoicechatClient.CLIENT_CONFIG.microphone.get())) {
@@ -41,7 +44,7 @@ public class SelectMicrophoneScreen extends ListScreen<String> {
             return;
         }
         int bw = 60;
-        Button b = addRenderableWidget(new Button(width / 2 - bw / 2, guiTop + 35, bw, 20, new TranslatableComponent("message.voicechat.select"), button -> {
+        Button b = addRenderableWidget(new Button(width / 2 - bw / 2, guiTop + 35, bw, 20, SELECT, button -> {
             VoicechatClient.CLIENT_CONFIG.microphone.set(currentElement).save();
             button.active = false;
             ClientVoicechat client = ClientManager.getClient();
@@ -59,8 +62,8 @@ public class SelectMicrophoneScreen extends ListScreen<String> {
         int titleWidth = font.width(title);
         font.draw(stack, title.getVisualOrderText(), (float) (guiLeft + (xSize - titleWidth) / 2), guiTop + 7, getFontColor());
 
-        MutableComponent name = getCurrentElement() == null ? new TranslatableComponent("message.voicechat.no_microphone") : new TextComponent(SoundManager.cleanDeviceName(getCurrentElement()));
+        Component name = getCurrentElement() == null ? NO_MICROPHONE : new TextComponent(SoundManager.cleanDeviceName(getCurrentElement())).withStyle(ChatFormatting.WHITE);
         int nameWidth = font.width(name);
-        font.draw(stack, name.withStyle(ChatFormatting.WHITE).getVisualOrderText(), (float) (guiLeft + (xSize - nameWidth) / 2), guiTop + 7 + font.lineHeight + 7, 0);
+        font.draw(stack, name.getVisualOrderText(), (float) (guiLeft + (xSize - nameWidth) / 2), guiTop + 7 + font.lineHeight + 7, 0);
     }
 }