+++ common/src/main/java/de/maxhenkel/voicechat/gui/VoiceChatSettingsScreen.java
@@ -5,7 +5,7 @@ import com.mojang.blaze3d.vertex.PoseStack;
 import de.maxhenkel.voicechat.Voicechat;
 import de.maxhenkel.voicechat.VoicechatClient;
 import de.maxhenkel.voicechat.gui.widgets.*;
-import de.maxhenkel.voicechat.voice.client.ClientManager;
+import de.maxhenkel.voicechat.gui.volume.PlayerVolumesScreen;
 import de.maxhenkel.voicechat.voice.client.Denoiser;
 import de.maxhenkel.voicechat.voice.common.Utils;
 import net.minecraft.client.gui.components.Button;
@@ -20,6 +20,13 @@ import javax.annotation.Nullable;
 public class VoiceChatSettingsScreen extends VoiceChatScreenBase implements MicTestButton.MicListener {
 
     private static final ResourceLocation TEXTURE = new ResourceLocation(Voicechat.MODID, "textures/gui/gui_voicechat_settings.png");
+    private static final Component TITLE = new TranslatableComponent("gui.voicechat.voice_chat_settings.title");
+    private static final Component ENABLED = new TranslatableComponent("message.voicechat.enabled");
+    private static final Component DISABLED = new TranslatableComponent("message.voicechat.disabled");
+    private static final Component ADJUST_VOLUMES = new TranslatableComponent("message.voicechat.adjust_volumes");
+    private static final Component SELECT_MICROPHONE = new TranslatableComponent("message.voicechat.select_microphone");
+    private static final Component SELECT_SPEAKER = new TranslatableComponent("message.voicechat.select_speaker");
+    private static final Component BACK = new TranslatableComponent("message.voicechat.back");
 
     private double micValue;
     private VoiceActivationSlider voiceActivationSlider;
@@ -28,7 +35,7 @@ public class VoiceChatSettingsScreen extends VoiceChatScreenBase implements MicT
     private Screen parent;
 
     public VoiceChatSettingsScreen(@Nullable Screen parent) {
-        super(new TranslatableComponent("gui.voicechat.voice_chat_settings.title"), 248, 219);
+        super(TITLE, 248, 219);
         this.parent = parent;
     }
 
@@ -47,9 +54,7 @@ public class VoiceChatSettingsScreen extends VoiceChatScreenBase implements MicT
         addRenderableWidget(new MicAmplificationSlider(guiLeft + 10, y, xSize - 20, 20));
         y += 21;
         BooleanConfigButton denoiser = addRenderableWidget(new BooleanConfigButton(guiLeft + 10, y, xSize - 20, 20, VoicechatClient.CLIENT_CONFIG.denoiser, enabled -> {
-            return new TranslatableComponent("message.voicechat.denoiser",
-                    enabled ? new TranslatableComponent("message.voicechat.enabled") : new TranslatableComponent("message.voicechat.disabled")
-            );
+            return new TranslatableComponent("message.voicechat.denoiser", enabled ? ENABLED : DISABLED);
         }));
         if (Denoiser.createDenoiser() == null) {
             denoiser.active = false;
@@ -68,20 +73,20 @@ public class VoiceChatSettingsScreen extends VoiceChatScreenBase implements MicT
         addRenderableWidget(new MicTestButton(guiLeft + 10, y, xSize - 20, 20, this));
         y += 21;
         if (isIngame()) {
-            addRenderableWidget(new Button(guiLeft + 10, y, xSize - 20, 20, new TranslatableComponent("message.voicechat.adjust_volumes"), button -> {
-                minecraft.setScreen(new AdjustVolumeScreen(this, ClientManager.getPlayerStateManager().getPlayerStates(false)));
+            addRenderableWidget(new Button(guiLeft + 10, y, xSize - 20, 20, ADJUST_VOLUMES, button -> {
+                minecraft.setScreen(new PlayerVolumesScreen());
             }));
             y += 21;
         }
-        addRenderableWidget(new Button(guiLeft + 10, y, xSize / 2 - 15, 20, new TranslatableComponent("message.voicechat.select_microphone"), button -> {
+        addRenderableWidget(new Button(guiLeft + 10, y, xSize / 2 - 15, 20, SELECT_MICROPHONE, button -> {
             minecraft.setScreen(new SelectMicrophoneScreen(this));
         }));
-        addRenderableWidget(new Button(guiLeft + xSize / 2 + 6, y, xSize / 2 - 15, 20, new TranslatableComponent("message.voicechat.select_speaker"), button -> {
+        addRenderableWidget(new Button(guiLeft + xSize / 2 + 6, y, xSize / 2 - 15, 20, SELECT_SPEAKER, button -> {
             minecraft.setScreen(new SelectSpeakerScreen(this));
         }));
         y += 21;
         if (!isIngame() && parent != null) {
-            addRenderableWidget(new Button(guiLeft + 10, y, xSize - 20, 20, new TranslatableComponent("message.voicechat.back"), button -> {
+            addRenderableWidget(new Button(guiLeft + 10, y, xSize - 20, 20, BACK, button -> {
                 minecraft.setScreen(parent);
             }));
         }
@@ -106,9 +111,8 @@ public class VoiceChatSettingsScreen extends VoiceChatScreenBase implements MicT
 
     @Override
     public void renderForeground(PoseStack poseStack, int mouseX, int mouseY, float delta) {
-        Component title = new TranslatableComponent("gui.voicechat.voice_chat_settings.title");
-        int titleWidth = font.width(title);
-        font.draw(poseStack, title.getVisualOrderText(), (float) (guiLeft + (xSize - titleWidth) / 2), guiTop + 7, getFontColor());
+        int titleWidth = font.width(TITLE);
+        font.draw(poseStack, TITLE.getVisualOrderText(), (float) (guiLeft + (xSize - titleWidth) / 2), guiTop + 7, getFontColor());
     }
 
     @Override